pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "atari-dqn"
version = "0.1.0"
description = "Modular Atari DQN implementation with paper-faithful and modern tracks"
readme = "final_implementation_plan.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "numpy>=1.26.0,<2.0",
    "gymnasium[atari]>=1.0.0",
    "ale-py>=0.10.0",
    "shimmy[atari]>=2.0.0",
    "opencv-python-headless>=4.9.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.17.0",
    "imageio[ffmpeg]>=2.34.0",
    "tqdm>=4.66.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dashboard = ["streamlit>=1.40.0", "plotly>=5.24.0", "pandas>=2.2.0"]
dev = ["pytest>=8.0.0", "pytest-cov>=5.0.0", "ruff>=0.5.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/dqn"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
